Rename rmui codemod and include ThemeProviderAdder in the same executable
